skelton/requirements.txt
aws-sam-translator==1.28.1
aws-utils<=0.1
boto3==1.28.64
botocore>=1.31.64
flask==3.0.0
fsspec==2023.10.0
joblib==1.3.2
numpy==1.24.4
pandas==2.0.3
pre-commit<=3.6.0
pymsteams==0.1.9
PyYAML==6.0
requests==2.31.0
s3fs==2023.10.0
sagemaker==2.197.0
scikit-learn==1.1
stepfunctions==2.3.0
tabulate==0.9.0
